# backend/services-v2/identity-service/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Service API
  version: 2.0.0
  description: Authentication and authorization service for Hospital Management System V2
  contact:
    name: Hospital Management Team
  license:
    name: MIT
servers:
  - url: http://localhost:3021
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register patient
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName]
              properties:
                email: {type: string, format: email}
                password: {type: string, minLength: 8}
                fullName: {type: string}
                phoneNumber: {type: string}
                citizenId: {type: string}
                dateOfBirth: {type: string, format: date}
                gender: {type: string, enum: [male, female, other]}
                address: {type: string}
      responses:
        '201':
          description: Created
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: {type: string}
                password: {type: string}
                mfaCode: {type: string}
      responses:
        '200':
          description: OK
...
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: {type: string}
      responses:
        '200':
          description: OK
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Forgot password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: {type: string}
      responses:
        '200':
          description: OK
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessToken, refreshToken, newPassword]
              properties:
                accessToken: {type: string}
                refreshToken: {type: string}
                newPassword: {type: string}
      responses:
        '200':
          description: OK
  /auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
      responses:
        '200':
          description: OK

  # User Management
...
  /auth/mfa/enable:
    post:
      tags: [MFA]
      summary: Enable MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method: {type: string, enum: [totp, sms, email]}
      responses:
        '200':
          description: OK
  /auth/mfa/verify:
    post:
      tags: [MFA]
      summary: Verify MFA code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: {type: string}
      responses:
        '200':
          description: OK
  /auth/mfa/disable:
    post:
      tags: [MFA]
      summary: Disable MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: {type: string}
      responses:
        '200':
          description: OK

  # Session Management
...
  /api/v1/password-policy:
    get:
      tags: [Password Policy]
      summary: Get password policy
      security: []
      responses:
        '200':
          description: OK
    put:
      tags: [Password Policy]
      summary: Update password policy (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                minLength: {type: integer}
                requireUppercase: {type: boolean}
                requireLowercase: {type: boolean}
                requireNumbers: {type: boolean}
                requireSpecialChars: {type: boolean}
      responses:
        '200':
          description: OK
  /api/v1/password-policy/validate:
    post:
      tags: [Password Policy]
      summary: Validate password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: {type: string}
      responses:
        '200':
          description: OK

  # Account Recovery
  /api/v1/account-recovery/methods:
    get:
      tags: [Account Recovery]
      summary: Get recovery methods
      responses:
        '200':
          description: OK
    put:
      tags: [Account Recovery]
      summary: Update recovery methods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recoveryEmail: {type: string}
      responses:
        '200':
          description: OK
  /api/v1/account-recovery/request-reset:
    post:
      tags: [Account Recovery]
      summary: Request password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: {type: string}
                method: {type: string}
      responses:
        '200':
          description: OK
  /api/v1/account-recovery/verify-token:
    post:
      tags: [Account Recovery]
      summary: Verify reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
      responses:
        '200':
          description: OK
  /api/v1/account-recovery/reset-password:
    post:
      tags: [Account Recovery]
      summary: Reset password with token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword, confirmPassword]
              properties:
                token: {type: string}
                newPassword: {type: string}
                confirmPassword: {type: string}
      responses:
        '200':
          description: OK
  /api/v1/account-recovery/history:
    get:
      tags: [Account Recovery]
      summary: Get recovery history
      parameters:
        - name: page
          in: query
          schema: {type: integer}
        - name: pageSize
          in: query
          schema: {type: integer}
      responses:
        '200':
          description: OK

  # Permission Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT